Task Details

+++ task-force-frontend/src/app/task/task-card/task-card.component.ts
@@ -1,6 +1,7 @@
-import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
+import { Component, OnInit, Input } from '@angular/core';
 import { Task } from 'src/app/task';
 import { TaskService } from 'src/app/task.service';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-task-card',
@@ -9,11 +10,11 @@ import { TaskService } from 'src/app/task.service';
 })
 export class TaskCardComponent implements OnInit {
   @Input() task$: Task = null;
-  @Output() notifyParentOfChange: EventEmitter<string> = new EventEmitter<string>();
+
   newTitle$: string = '';
   updated$: boolean = false;
 
-  constructor(private taskService: TaskService) { }
+  constructor(private taskService: TaskService, private router: Router) { }
 
   ngOnInit(): void {
     this.newTitle$ = this.task$.title;
@@ -44,7 +45,6 @@ export class TaskCardComponent implements OnInit {
 
   markComplete() {
     this.taskService.patchTask(this.task$.id).subscribe(() => {
-      this.notifyParentOfChange.emit('Marking Task Complete');
       console.log("Task marked completed successfully.");
     },
       err => {
@@ -55,7 +55,6 @@ export class TaskCardComponent implements OnInit {
   markUncomplete() {
     this.task$.completed = !this.task$.completed;
     this.taskService.updateTask(this.task$).subscribe(() => {
-      this.notifyParentOfChange.emit('Marking Task Uncomplete');
       console.log("Task marked uncompleted successfully.");
     },
       err => {
@@ -64,12 +63,14 @@ export class TaskCardComponent implements OnInit {
   }
 
   deleteTask() {
-    this.taskService.deleteTask(this.task$.id).subscribe(() => {
-      this.notifyParentOfChange.emit('Deleting Task');
-      console.log("Task deleted successfully.");
-    },
-      err => {
-        console.log(err);
-      });
+    if (confirm("Are you sure you want to delete this task?")) {
+      this.taskService.deleteTask(this.task$.id).subscribe(() => {
+        console.log("Task deleted successfully.");
+        this.router.navigate(['/', 'tasks']);
+      },
+        err => {
+          console.log(err);
+        });
+    }
   }
 }
